App flex-wrapper, PageNotFound renamed, footer updated
