+++ src/App.scss
@@ -1,1 +1,6 @@
-// empty
+.flex-wrapper {
+  display: flex;
+  min-height: 100vh;
+  flex-direction: column;
+  justify-content: space-between;
+}
