+++ src/index.tsx
@@ -13,6 +13,5 @@ root.render(
     <Router>
       <App />
     </Router>
-    ,
   </React.StrictMode>,
 );
