--- src/pages/PageNotFound/index.ts
@@ -1,1 +1,0 @@
-export * from './PageNotFound';
